Changed method in Vault Ext
AllowedVirtualNetworks() arguments changed from:
(params (Input<string> subnetId, Input<bool> ignoreMissingVnetServiceEndpoint)[] subnets)

The bool argument in the tuple seemed unfriendly and unnecessary to be a required argument, as the ignoreMissingVnetServiceEndpoint is not a very commonly used or specified property for Vault.

Changed to:
(Input<string> subnetId, bool ignoreMissingVnetServiceEndpoint = false)

this takes two arguments, though the bool is optional and defaulted as false.
For the user to add multiple subnets to the whitelist, the method must be called multiple times, passing in the subnet ID each time.